tests: modify tr expressions so that xpg4/tr handles it on Solaris
It seems that xpg4/tr mishandles some strings involving [ not followed
by a character class:

    % echo '[::1]' | /usr/xpg4/bin/tr -d '[]'
    [::1

    % echo '[::1]' | /usr/xpg4/bin/tr -d '['
    usr/xpg4/bin/tr: Bad string.

This was breaking two tests. To fix the issue, use the octal
representations of [ and ] instead. Reference the octal expression
as a newly exported variable so that it's shared across tests and
more easily understood when reading it.

Signed-off-by: Ben Walton <bdwalton@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>