tests: fix sed usage in tests to work around broken xpg4/sed on Solaris
In 99094a7a (t: fix trivial &&-chain breakage, 2015-03-20), a
trivial && breakage was fixed.  This exposed a problem with the test
when run on Solaris with xpg4/sed that had gone silently undetected
since its introduction in e4bd10b2. Solaris' sed executes the
requested substitution but prints a warning about the missing
newline at the end of the file and exits with status 2.

    % echo "CHANGE_ME" | \
    tr -d "\\012" | /usr/xpg4/bin/sed -e 's/CHANGE_ME/change_me/'
    sed: Missing newline at end of file standard input.
    change_me
    % echo $?
    2

To work around this, use perl to perform the substitution instead.

Signed-off-by: Ben Walton <bdwalton@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>