builtin/mv: remove get_pathspec()
`get_pathspec` is deprecated and builtin/mv.c is its last caller, so
getting rid of `get_pathspec` is rather easy.

By getting rid of `get_pathspec`, the documentation such as
'technical/api-setup.txt' becomes easier to read as the reader
doesn't need to bear with the additional fact that `get_pathspec` is
deprecated.

The code in builtin/mv.c still requires some work to make it less
ugly.

Signed-off-by: Stefan Beller <sbeller@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>